requirements.txt
python-telegram-bot==22.1
openai==1.35.7
python-dotenv==1.0.1
aiosqlite==0.21.0
Pillow==11.2.1